+++ src/interfaces/connector-options.ts
@@ -13,7 +13,7 @@ export interface ShogunConnectorOptions {
   showMetamask?: boolean;
   showWebauthn?: boolean;
   showNostr?: boolean;
-  showOauth?: boolean;
+  showZkProof?: boolean;
   darkMode?: boolean;
 
   gunInstance?: IGunInstance;
@@ -25,16 +25,6 @@ export interface ShogunConnectorOptions {
     signup?: number;
     operation?: number;
   };
-  oauth?: {
-    providers: Record<
-      string,
-      {
-        clientId: string;
-        clientSecret?: string;
-        redirectUri?: string;
-      }
-    >;
-  };
   webauthn?: {
     enabled?: boolean;
   };
@@ -43,6 +33,10 @@ export interface ShogunConnectorOptions {
   };
   web3?: {
     enabled?: boolean;
+  };
+  zkproof?: {
+    enabled?: boolean;
+    defaultGroupId?: string;
   };
   enableGunDebug?: boolean;
   enableConnectionMonitoring?: boolean;